Make identity and last function tests pass
I wrote code to make the identity function (trivial) and the
last function pass. Last took a couple of iterations. I wanted to
format it using the unpopular ? operator like in the example of the
first function, but I couldn't think of a way to do it without
embedding a second if clause, and it started looking ugly.