Make reject function return arrays - not to spec
This commit saves my closest attempt yet to the preferred approach
to the _.reject function. I've got it returning arrays, and it returns
differing results for the even & odd tests, but it still doesn't
actually return the correct results yet. I'm having real trouble
understanding how to correctly call the _.filter function, and
specifically how to deal with the test function.