Another attempt to resolve inscrutable function syntax
I'm having a really hard time with this one. The layers of function call
syntax gives me a lot of trouble. I think what's making it really hard is
that the test function needs to stay abstracted. Anyway... I'm moving on
and I will do the copy-pasta method on my next commit, just to be able to
finish the remainder of the lesson.