Remove comments and some unnecessary stuff from reduce
Cleaned up reduce a little more, removing the old commented-out code
that I was using before. Also felt it was unnecessary to make an explicit
copy of the last n-1 elements of the array.

Important point about this version of the reduce function: it only
handles arrays, not objects now.