Refactor _.every with lessons learned from _.contains
So I finally looked at the valuable resource I was given: the example
of the _.contains function, which I had somehow glossed over before,
and who'da thunk, it turns out that it very instructive when I spent
time actually reasoning out what it was doing in there.

I need to be paying more attention to what's going on in the argument
structure of the function calls in functions that I am reusing. That
is the TL;DR.