Implement _.memoize and _.delay and _.shuffle functions
Oops! I forgot to be committing in between some of these functions. I still
don't have 100% test case coverage on memoize, so I might come back to it
later. But unfortunately I'm running out of time to work on the Twittler
assignment, so perfection will have to wait, for now.